+++ pyforestscan_qgis/worker/__init__.py
@@ -1,0 +1,1 @@
+"""External worker entrypoints for PyForestScan batch execution."""